requirements.txt
Django==4.1.7
openai==0.27.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
email-validator==1.3.1